pyproject.toml
[tool.poetry]
name = "nlp"
version = "0.1.0"
description = ""
authors = ["eng-alecrim <lucasalecrimvieira@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
wordcloud = "^1.9.3"
matplotlib = "^3.9.0"
scikit-learn = "^1.5.1"
spacy = "^3.7.5"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.0"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.6"
ignr = "^2.2"
black = "^24.8.0"
jupyterlab = "^4.2.4"
catppuccin-jupyterlab = "^0.2.3"

[tool.ruff]
line-length = 79
extend-exclude = ["migrations"]

[tool.ruff.lint]
preview = true
select = ["I", "F", "E", "W", "PL", "PT"]
ignore = ["E501", "E731", "PLR1711"]

[tool.ruff.format]
indent-style = "space"
preview = true
quote-style = "double"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "-p no:warnings"

[tool.taskipy.tasks]
lint = "ruff check . && ruff check . --diff"
format = "ruff check . --fix && ruff format ."
pre_test = "task lint"
test = "pytest -s -x --cov=nlp -vv"
post_test = "coverage html"
jupyter = "python -m jupyterlab"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
